requirements-test.txt
# Core test extras from setup.py
plone.app.testing
plone.app.robotframework[debug]
pytest
pytest-cov
pytest-plone
shutup

# (Optional) Lock specific versions if needed
# pytest==7.4.4
# pytest-cov==4.1.0

# Install your package and test extras in editable mode
-e .[test]